Title: 올바른 등식 만들기, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240518/올바른 등식 만들기/right-equality.cpp
@@ -6,7 +6,7 @@ int n, m;
 
 int OFFSET = 30;
 int numbers[105];
-int dp[60][105];
+long long dp[60][105];
 
 int main() {
